# Conflicts:
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$1.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$10.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$11.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$12.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$13.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$14.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$15.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$16.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$17.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$18.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$19.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$2.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$20.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$21.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$22.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$23.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$24.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$25.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$26.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$27.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$28.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$29.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$3.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$30.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$31.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$32.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$33.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$34.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$35.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$36.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$37.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$38.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$39.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$4.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$40.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$41.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$42.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$43.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$44.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$45.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$46.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$47.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$48.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$49.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$5.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$50.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$51.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$52.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$53.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$6.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$7.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$8.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView$9.class
#	oberdanEquipamientos/build/classes/Views/ABMClientesView.class

+++ oberdanEquipamientos/src/Views/JD_Producto_Nuevo.java
@@ -15,6 +15,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import javax.swing.JFrame;
 import javax.swing.JOptionPane;
 
 /**
@@ -23,6 +24,7 @@ import javax.swing.JOptionPane;
  */
 public class JD_Producto_Nuevo extends javax.swing.JDialog{
     private ProductoDAO productoDAO;
+    private ProductosView view;
     private List<Integer> list;
     int id_proveedor_actual = -1;
     
@@ -35,6 +37,14 @@ public class JD_Producto_Nuevo extends javax.swing.JDialog{
         productoDAO = ProductoDAO.getInstance();
         list = new ArrayList<>();
        
+    }
+
+    JD_Producto_Nuevo(JFrame jFrame, boolean b, ProductosView aThis) {
+        super(jFrame, b);
+        initComponents();
+        productoDAO = ProductoDAO.getInstance();
+        list = new ArrayList<>();
+        view = aThis;
     }
 
     /**
@@ -226,22 +236,22 @@ public class JD_Producto_Nuevo extends javax.swing.JDialog{
                 else
                     p.setStockMin(0);
                 list.add(id_proveedor_actual);
-                ResultSet rs = productoDAO.productoEliminado(p.getId());
-                try {
-                    rs.next();
-                    int i = 0;
-                    int dialogButton = JOptionPane.YES_NO_OPTION;
-                    int coincidencia = rs.getInt("producto_id");
-                    System.out.println("coincidencia = " + coincidencia);
-                    int resp = JOptionPane.showConfirmDialog(null,"El codigo "+p.getId()+" pertenece a un producto ingresado.\n ¿ esta seguro que desea reactivarlo ?","Importante",dialogButton);
-                    if(JOptionPane.YES_OPTION == resp){
-                        productoDAO.eliminarProducto(p);
-                        productoDAO.actualizarProducto(p);
-                        dispose();
+                int rs = productoDAO.productoEliminado(p.getId());
+                    if(rs==-1){
+                        productoDAO.nuevoProducto(p);
+                        
+                    }else{
+                        int i = 0;
+                        int dialogButton = JOptionPane.YES_NO_OPTION;
+                        int coincidencia = rs;
+                        int resp = JOptionPane.showConfirmDialog(null,"El codigo "+p.getId()+" pertenece a un producto ingresado.\n ¿ esta seguro que desea reactivarlo ?","Importante",dialogButton);
+                        if(JOptionPane.YES_OPTION == resp){
+                            productoDAO.eliminarProducto(p);
+                            productoDAO.actualizarProducto(p);
+                            dispose();
+                        }
                     }
-                } catch (SQLException ex) {
-                    productoDAO.nuevoProducto(p, list);
-                }
+                    
             }else
                 JOptionPane.showMessageDialog(null,"Debe ingresar un nombre y código del producto (valor numérico)", 
                             "Error",  JOptionPane.ERROR_MESSAGE);
